Pass context objects to modules on import
Changed the interpreter and transpiler to pass module context information when the module is loaded.
