+++ src/createContext.ts
@@ -13,7 +13,7 @@ import * as parser from './stdlib/parser'
 import * as stream from './stdlib/stream'
 import { streamPrelude } from './stdlib/stream.prelude'
 import { createTypeEnvironment, tForAll, tVar } from './typeChecker/typeChecker'
-import { Context, CustomBuiltIns, Environment, NativeStorage, Value, Variant } from './types'
+import { Context, CustomBuiltIns, Environment, ModuleContext, NativeStorage, Value, Variant } from './types'
 import { makeWrapper } from './utils/makeWrapper'
 import * as operators from './utils/operators'
 import { stringify } from './utils/stringify'
@@ -128,14 +128,18 @@ export const createEmptyContext = <T>(
   variant: Variant = 'default',
   externalSymbols: string[],
   externalContext?: T,
-  moduleParams?: any
+  moduleParams?: Map<string, any>
 ): Context<T> => {
+
+  if (moduleParams == null) {
+    moduleParams = new Map<string, any>();
+  }
+
   return {
     chapter,
     externalSymbols,
     errors: [],
     externalContext,
-    moduleParams,
     runtime: createEmptyRuntime(),
     numberOfOuterEnvironments: 1,
     prelude: null,
@@ -143,6 +147,8 @@ export const createEmptyContext = <T>(
     nativeStorage: createNativeStorage(),
     executionMethod: 'auto',
     variant,
+    moduleParams,
+    moduleContexts: new Map<string, ModuleContext>(),
     unTypecheckedCode: [],
     typeEnvironment: createTypeEnvironment(chapter),
     previousCode: []
