+++ src/transpiler/evalContainer.ts
@@ -1,7 +1,8 @@
-import { MODULE_PARAMS_ID, NATIVE_STORAGE_ID } from '../constants'
+import { ModuleState } from '..'
+import { MODULE_PARAMS_ID, MODULE_STATE_ID, NATIVE_STORAGE_ID } from '../constants'
 import { NativeStorage } from '../types'
 
-type Evaler = (code: string, nativeStorage: NativeStorage, moduleParams: any) => any
+type Evaler = (code: string, nativeStorage: NativeStorage, moduleParams: any, moduleState: ModuleState) => any
 
 /*
   We need to use new Function here to ensure that the parameter names do not get
@@ -12,6 +13,7 @@ export const sandboxedEval: Evaler = new Function(
   'code',
   NATIVE_STORAGE_ID,
   MODULE_PARAMS_ID,
+  MODULE_STATE_ID,
   `
   if (${NATIVE_STORAGE_ID}.evaller === null) {
     return eval(code);
