+++ src/transpiler/transpiler.ts
@@ -2,7 +2,7 @@ import { generate } from 'astring'
 import * as es from 'estree'
 import { SourceMapGenerator } from 'source-map'
 
-import { MODULE_PARAMS_ID, NATIVE_STORAGE_ID } from '../constants'
+import { MODULE_PARAMS_ID, MODULE_STATE_ID, NATIVE_STORAGE_ID } from '../constants'
 import { UndefinedVariable } from '../errors/errors'
 import { memoizedGetModuleFile } from '../modules/moduleLoader'
 import { AllowedDeclarations, Context, NativeStorage } from '../types'
@@ -47,7 +47,7 @@ function prefixModule(program: es.Program): string {
     prefix += `const __MODULE_${moduleCounter}__ = (${moduleText.substring(
       0,
       moduleText.length - 1
-    )})(${MODULE_PARAMS_ID});\n`
+    )})(${MODULE_PARAMS_ID}, ${MODULE_STATE_ID});\n`
     moduleCounter++
   }
   return prefix
